+++ clients/fiat/src/rest-api/rest-api.ts
@@ -15,13 +15,17 @@ import { ConfigurationRestAPI, RestApiResponse, sendRequest } from '@binance/com
 import { FiatApi } from './modules/fiat-api';
 
 import type {
+    FiatWithdrawRequest,
     GetFiatDepositWithdrawHistoryRequest,
     GetFiatPaymentsHistoryRequest,
+    GetOrderDetailRequest,
 } from './modules/fiat-api';
 
 import type {
+    FiatWithdrawResponse,
     GetFiatDepositWithdrawHistoryResponse,
     GetFiatPaymentsHistoryResponse,
+    GetOrderDetailResponse,
 } from './types';
 
 export class RestAPI {
@@ -68,6 +72,30 @@ export class RestAPI {
     }
 
     /**
+     * Submit withdraw request, in this version, we only support BRL withdrawal via bank_transfer.
+     *
+     * You need to call this api first, and call query order detail api in a loop to get the status of the order until this order is successful.
+     *
+     * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
+     *
+     * you need to bind your bank account on web/app before using the corresponding account number
+     *
+     * Weight: 45000
+     *
+     * @summary Fiat Withdraw(WITHDRAW)
+     * @param {FiatWithdrawRequest} requestParameters Request parameters.
+     *
+     * @returns {Promise<RestApiResponse<FiatWithdrawResponse>>}
+     * @throws {RequiredError | ConnectorClientError | UnauthorizedError | ForbiddenError | TooManyRequestsError | RateLimitBanError | ServerError | NotFoundError | NetworkError | BadRequestError}
+     * @see {@link https://developers.binance.com/docs/fiat/rest-api/Fiat-Withdraw Binance API Documentation}
+     */
+    fiatWithdraw(
+        requestParameters: FiatWithdrawRequest = {}
+    ): Promise<RestApiResponse<FiatWithdrawResponse>> {
+        return this.fiatApi.fiatWithdraw(requestParameters);
+    }
+
+    /**
      * Get Fiat Deposit/Withdraw History
      *
      * If beginTime and endTime are not sent, the recent 30-day data will be returned.
@@ -110,5 +138,25 @@ export class RestAPI {
         requestParameters: GetFiatPaymentsHistoryRequest
     ): Promise<RestApiResponse<GetFiatPaymentsHistoryResponse>> {
         return this.fiatApi.getFiatPaymentsHistory(requestParameters);
+    }
+
+    /**
+     * Get Order Detail
+     *
+     * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
+     *
+     * Weight: 45000
+     *
+     * @summary Get Order Detail(USER_DATA)
+     * @param {GetOrderDetailRequest} requestParameters Request parameters.
+     *
+     * @returns {Promise<RestApiResponse<GetOrderDetailResponse>>}
+     * @throws {RequiredError | ConnectorClientError | UnauthorizedError | ForbiddenError | TooManyRequestsError | RateLimitBanError | ServerError | NotFoundError | NetworkError | BadRequestError}
+     * @see {@link https://developers.binance.com/docs/fiat/rest-api/Get-Order-Detail Binance API Documentation}
+     */
+    getOrderDetail(
+        requestParameters: GetOrderDetailRequest
+    ): Promise<RestApiResponse<GetOrderDetailResponse>> {
+        return this.fiatApi.getOrderDetail(requestParameters);
     }
 }